Initial commit from code that I've altered from Silke's repository.
This commit has all of the requisite files to compute rigid clusters that I took from Silke's repository which is linked in the README.  suspensionRigidCluster.py is the only piece of code that we've written and every other file is a copy of Silke's code or alterations of her code.

So far, the only alterations we have made to her code are in Configuration_LFDEM.py.  There I have changed the the method readSimdata inside the Configuration class.  Instead of reading simulation data from a text file, the simulation data is just fed in as a function argument.

suspensionRigidCluster.py contains a single function that is supposed to take in output files from LF_DEM and feed them into Silke's rigid cluster code.