Removed the dataFile input.  Changed code so it excludes frictionless contacts.
As Toka and Abhi I discussed yesterday we realized that we don't need the dataFile input since it only provides the strain rates for each timestep which aren't actually needed.  I also changed part of the contact finder code so it only looks at particles that are at or above the Coloumb threshold.