Update rigidClusterProcessor.py
Recently I found out that the rigid_cluster function in Pebbles.py is not working correctly.  It seems that it is miscounting the number of rigid clusters and the number of bonds they have.  Instead of using the output files from the rigid_cluster function I extracted all the needed information from the pebbles object directly.  Assuming the error in rigid_cluster is contained to just it's main output variables, this should fix the problem.  Waiting on Silke to get back to me before I implement a fix inside the Pebbles.py rigid_cluster function.