Finished first version of the LF_DEM hessian generator.
This new file "hessianNew.py" contains three functions with the last one, "generateHessianFiles.py" being the most important.  It takes in a directory and finds all the par, int, and data files in that directory and then uses those to generate a 2D frictional hessian like the one that Silke made for frictional packings.